% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{influridge}
\alias{influridge}
\title{Influence of single observations on the choice of the penalty parameter in ridge regression}
\usage{
influridge(
  X,
  y,
  nw = 100,
  max.weight = 4,
  noExpand = 0,
  noShrink = 0,
  degreeFreedom = FALSE,
  control.list = list(factr = 1e-08)
)
}
\arguments{
\item{X}{numerical input matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{numerical output vector; response variable.}

\item{nw}{scalar, number of grid points for curve.}

\item{max.weight}{scalar, the maximum weight for the curve.}

\item{noExpand}{scalar, number of expanders to be highlighted (less than 5).}

\item{noShrink}{scalar, number of shrinkers to be highlighted (less than 5).}

\item{degreeFreedom}{logical, should the degrees of freedom be plotted instead of the tuning parameter (default = FALSE).}

\item{control.list}{list, a list of control parameters for the optim function. See 'Details' under graphics::optim.}
}
\description{
This function produces curve plot showing the optimal tuning paramer when up- or downweighting each observation
}
\examples{

p <- 5 
n <- 50 
sigma <- 1 
beta <- rep(0.1,p)

Simulating design matrix, X
setseed(556)
X <- matrix(rnorm(n*p),n,p)

Simulate outcome vector, Y
y <- X \%*\% beta + rnorm(n, 0, sigma)

Plot curves (no highlighted shrinkers/expanders)
influridge(X,y)

Simulate outcome vector Y, adding a large negative residual to observation 7.
y <- X \%*\% beta + c(rnorm(6, 0, sigma), -4, rnorm(n - 7, 0, sigma))
influridge(X,y,noShrink = 1)

Plot degrees of freedom
y <- X \%*\% beta + c(rnorm(6, 0, sigma), -4, rnorm(n - 7, 0, sigma))
influridge(X,y,noShrink = 1,degreeFreedom == TRUE)

}
