% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_influridge}
\alias{plot_influridge}
\title{Influence of single observations on the choice of the penalty parameter in ridge regression}
\usage{
plot_influridge(
  X,
  y,
  nw = 100,
  max.weight = 4,
  control.list = list(factr = 1e-04)
)
}
\arguments{
\item{X}{numerical input matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{numerical output vector; response variable.}

\item{nw}{scalar, number of grid points for curve.}

\item{max.weight}{scalar, the maximum weight for the curve.}
}
\description{
This function produces curve plot showing the optimal tuning paramer when up- or downweighting each observation
}
