% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuning.cv.svd}
\alias{tuning.cv.svd}
\title{Weighted leave-one-out crossvalidation error}
\usage{
tuning.cv.svd(lambda, w, svd.int, y.int)
}
\arguments{
\item{lambda}{scalar, ridge tuning parameter.}

\item{w}{vector of length n, weigths for all observations.}

\item{svd.int}{svd object, singular value decomposition of the input matrix.}

\item{y.int}{numerical output vector; response variable.}
}
\description{
This function produces the leave-one-out squared cross-validation error for a given tuning parameter and weight.
}
